Refactor: move unit test for Config to standalone file
